Edit seed reconciliation flow
